+++ library/UnitTests/Collections/TestSegCollection.cpp
@@ -1627,4 +1627,5 @@ namespace Nektar
             }
         }
     }
+
 }